Init of temporary branch Implementation of image conversion script All the content has to be migrated to dev_toni
Branch contains working script that converts pixelart to matrix usable in rviz occupancy grid
Branch contains code that has to be implemented to launchfile to reference the method declared in the scripts

Code is very rudimentary but works, a lot of the origninal launchscripts were destroyed, branch must not be merged. Implementation has to be done manually.

This branch was only for experimenting purposes